requirements.txt
pandas>=2.1
numpy>=1.26
requests>=2.31
python-dotenv>=1.0